# Conflicts:
#	pom.xml
#	redisson-all/pom.xml
#	redisson-tomcat/pom.xml
#	redisson-tomcat/redisson-tomcat-6/pom.xml
#	redisson-tomcat/redisson-tomcat-7/pom.xml
#	redisson-tomcat/redisson-tomcat-8/pom.xml
#	redisson/pom.xml

+++ redisson/src/test/java/org/redisson/BaseTest.java
@@ -38,7 +38,9 @@ public abstract class BaseTest {
             if (redisson == null) {
                 redisson = defaultRedisson;
             }
-            redisson.getKeys().flushall();
+            if (flushBetweenTests()) {
+                redisson.getKeys().flushall();
+            }
         }
     }
 
@@ -74,4 +76,7 @@ public abstract class BaseTest {
         return Redisson.create(config);
     }
 
+    protected boolean flushBetweenTests() {
+        return true;
+    }
 }
